Circle CI: Adds timeout to maven build
Sometimes the job lands on a weird VM where the
maven build hangs for many minutes with no output
to the log, e.g. 10 minutes in:

```Resolving 216 out of 524 artifacts```

If there is no output to the log for such a long time,
Cirlce CI terminates the job. To avoid these instabilities
that have nothing to do with our code,
I entered a biger timeout. It seems to fix the problem
according to the last trial.

Signed-off-by: Michal Karm Babacek <karm@fedoraproject.org>