Quarkus 1.7.0.Final, JWT lib token generation to Vert.x JWT 3.9.2
This patch modernizes Quarkus generator to be on par with
what is deemed valid for Quarkus 1.7.0.final. Both JVM and Native
modes work. It also ports JWT example code to use Vert.x JWT 3.9.2.

All runtimes run with it. The only test failure recorded was
Liberty's mysterious error message unrelated to this patch (occured before),
which is tracked here: issues/328 #328

Signed-off-by: Michal Karm Babacek <karm@fedoraproject.org>